fixes Windows build
The fixed error is:

    src\with_read_timeout.rs:40:13: 40:19 error: unresolved import `select::fd_set`. Maybe a missing `extern crate select`? [E0432]
    src\with_read_timeout.rs:40         use select::fd_set;
                                            ^~~~~~
    src\with_read_timeout.rs:40:13: 40:19 help: run `rustc --explain E0432` to see a detailed explanation
    error: aborting due to previous error

Introduced in commit dfd91141a0c864bd939d9e0f9cd2ff578aeea2f8

Hint taken from
https://github.com/maidsafe/rust-utp/commit/c7a7190b53488cc1f132bc55a6686c3990370409#diff-b4aea3e418ccdb71239b96952d9cddb6R81
